Use byte for derived alphabet instead of char

+++ src/fm-index/Abstract/IPrefixSum.cs
@@ -2,6 +2,6 @@ namespace FmIndex.Abstract
 {
     internal interface IPrefixSum
     {
-        int this[int c] { get; }
+        int this[byte c] { get; }
     }
 }
